Extract chiban shared logic to 04_make_chiban_lib
